+++ src/main/java/ru/dpopkov/tasktracker/Main.java
@@ -1,11 +1,9 @@
 package ru.dpopkov.tasktracker;
 
 import ru.dpopkov.tasktracker.repositories.inmemory.SimpleProjectRepository;
+import ru.dpopkov.tasktracker.repositories.inmemory.SimpleTaskRepository;
 import ru.dpopkov.tasktracker.repositories.inmemory.SimpleUserRepository;
-import ru.dpopkov.tasktracker.services.ProjectService;
-import ru.dpopkov.tasktracker.services.ProjectServiceImpl;
-import ru.dpopkov.tasktracker.services.UserService;
-import ru.dpopkov.tasktracker.services.UserServiceImpl;
+import ru.dpopkov.tasktracker.services.*;
 import ru.dpopkov.tasktracker.ui.*;
 import ru.dpopkov.tasktracker.ui.actions.Actions;
 import ru.dpopkov.tasktracker.ui.actions.ActionPack;
@@ -21,7 +19,8 @@ public class Main {
     private static TaskTracker initTaskTracker() {
         ProjectService projectService = new ProjectServiceImpl(new SimpleProjectRepository());
         UserService userService = new UserServiceImpl(new SimpleUserRepository());
-        return new TaskTracker(userService, projectService);
+        TaskService taskService = new TaskServiceImpl(new SimpleTaskRepository());
+        return new TaskTracker(userService, projectService, taskService);
     }
 
     private static UiCycle initUiCycle(TaskTracker taskTracker) {