+++ src/main/java/ru/dpopkov/tasktracker/TaskTracker.java
@@ -1,10 +1,13 @@
 package ru.dpopkov.tasktracker;
 
 import ru.dpopkov.tasktracker.model.Project;
+import ru.dpopkov.tasktracker.model.Task;
 import ru.dpopkov.tasktracker.model.User;
 import ru.dpopkov.tasktracker.services.ProjectService;
+import ru.dpopkov.tasktracker.services.TaskService;
 import ru.dpopkov.tasktracker.services.UserService;
 
+import java.time.Duration;
 import java.util.List;
 import java.util.Optional;
 
@@ -12,10 +15,12 @@ public class TaskTracker {
 
     private final UserService userService;
     private final ProjectService projectService;
+    private final TaskService taskService;
 
-    public TaskTracker(UserService userService, ProjectService projectService) {
+    public TaskTracker(UserService userService, ProjectService projectService, TaskService taskService) {
         this.userService = userService;
         this.projectService = projectService;
+        this.taskService = taskService;
     }
 
     public void createUser(String firstName, String lastName) {
@@ -50,5 +55,22 @@ public class TaskTracker {
 
     public List<Project> getAllProjects() {
         return projectService.getAll();
+    }
+
+    public void createTask(String name, String description, Project project, Duration time) {
+        Task task = new Task(name, description, project, time);
+        taskService.create(task);
+    }
+
+    public Optional<Task> getTaskById(int id) {
+        return taskService.findById((long) id);
+    }
+
+    public boolean deleteTask(int id) {
+        return taskService.delete((long) id);
+    }
+
+    public List<Task> getAllTasks() {
+        return taskService.getAll();
     }
 }