+++ src/main/java/ru/dpopkov/tasktracker/repositories/inmemory/SimpleTaskRepository.java
@@ -1,6 +1,7 @@
 package ru.dpopkov.tasktracker.repositories.inmemory;
 
 import ru.dpopkov.tasktracker.model.Task;
+import ru.dpopkov.tasktracker.repositories.TaskRepository;
 
-public class SimpleTaskRepository extends SimpleBaseRepository<Task> {
+public class SimpleTaskRepository extends SimpleBaseRepository<Task> implements TaskRepository {
 }