+++ src/main/java/ru/dpopkov/tasktracker/services/TaskService.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.tasktracker.services;
+
+import ru.dpopkov.tasktracker.model.Task;
+
+public interface TaskService extends BaseService<Task> {
+}