+++ History.md
@@ -10,3 +10,4 @@ History of implementation
 7. Add i18n.
 8. Add operations to create, show and delete projects.
 9. Refactor - simplify inner classes in ActionPack.
+10. Add operations to create, show and delete tasks.